Install fpms.service with correct exec path
When packaging the source, prefix variable is changed, so we need to
correctly set ExecStart from fpms.service to the new path.

Debhelper dh_auto_build is not overriden with custom prefix, so we need
to update the fpms.service ExecStart prefix on install target.

Also copying oled-start directly to bin directory, instead of making a
symbolic link.

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>